requirements.txt
openai>=1.6.1
pydantic-ai>=0.0.3
requests==2.31.0
python-dotenv==1.0.0
sentence-transformers==2.2.2
pydantic>=2.5.2
colorama==0.4.6
numpy>=1.26.2
scikit-learn>=1.3.0
psycopg2-binary>=2.9.9
pytest==7.4.3
pytest-cov==4.1.0
